Split maven.go and pom.go.
